remove obsolete endpoints, adjust tests
