+++ dockstore-webservice/src/main/java/io/dockstore/webservice/resources/EntryResource.java
@@ -243,12 +243,12 @@ public class EntryResource implements AuthenticatedResourceInterface, AliasableR
     }
 
     @GET
-    @Path("/{alias}/aliases")
     @Timed
     @UnitOfWork(readOnly = true)
-    @Operation(operationId = "getEntryByAlias", description = "Retrieves an entry by alias.")
+    @Path("/{alias}/aliases")
+    @Operation(operationId = "getEntryByAlias", description = "Retrieves an entry by alias.", security = @SecurityRequirement(name = JWT_SECURITY_DEFINITION_NAME))
     @ApiResponse(responseCode = HttpStatus.SC_OK + "", description = "Successfully retrieved entry", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = Entry.class)))
-    public Entry getEntryByAlias(@Parameter(hidden = true, name = "user")@Auth Optional<User> user,
+    public Entry getEntryByAlias(@Parameter(hidden = true, name = "user") @Auth Optional<User> user,
             @Parameter(description = "Alias", name = "alias", in = ParameterIn.PATH, required = true) @PathParam("alias") String alias) {
         Entry<? extends Entry, ? extends Version> entry = toolDAO.getGenericEntryByAlias(alias);
         checkNotNullEntry(entry);
