+++ dockstore-webservice/src/main/resources/openapi3/openapi.yaml
@@ -1975,27 +1975,6 @@ paths:
       - BEARER: []
       tags:
       - containers
-  /containers/{alias}/aliases:
-    get:
-      description: Retrieves a tool by alias.
-      operationId: getToolByAlias
-      parameters:
-      - in: path
-        name: alias
-        required: true
-        schema:
-          type: string
-      responses:
-        default:
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/DockstoreTool'
-          description: default response
-      security:
-      - BEARER: []
-      tags:
-      - containers
   /containers/{containerId}:
     delete:
       description: Delete a tool.
@@ -2959,6 +2938,8 @@ paths:
               schema:
                 $ref: '#/components/schemas/Entry'
           description: Successfully retrieved entry
+      security:
+      - BEARER: []
       tags:
       - entries
   /entries/{entryId}/exportToOrcid:
@@ -7371,27 +7352,6 @@ paths:
                 type: array
                 items:
                   $ref: '#/components/schemas/WorkflowVersion'
-          description: default response
-      security:
-      - BEARER: []
-      tags:
-      - workflows
-  /workflows/{alias}/aliases:
-    get:
-      description: Retrieves a workflow by alias.
-      operationId: getWorkflowByAlias
-      parameters:
-      - in: path
-        name: alias
-        required: true
-        schema:
-          type: string
-      responses:
-        default:
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Workflow'
           description: default response
       security:
       - BEARER: []
